Fix DUT hang issue during wlan-recovery when matter is enabled.
[Description]
Fix DUT hang issue during wlan-recovery when matter is enabled.

[Root Cause]
The root cause of this issue is that during Wi-Fi recovery, wm_wifi.mcastf_mutex has not yet been initialized.
At this point, OT triggers a request for a UDP socket to join an IPv6 multicast group via the Wi-Fi netif,
which leads to a mutex acquisition failure and causes the system to get stuck.

[Fix]
The patch adds a check in the Wi-Fi driver’s mld_mac_filter function to verify whether the passed-in netif has already been added to the netif_list,
which indicates that wm_wifi and the initialization between the Wi-Fi driver and the network stack have been completed.

[Testing]
Passed

[Open Issues/Todo]
build pass

[Reviewers]
Gannicus Guo

[Port proposal]
general

Change-Id: 50238b52-b327-11f0-95d7-8b214fdeb63f
Signed-off-by: Maochen Wang <maochen.wang@nxp.com>