Add error check for TX RX histogram configuration
[Description]
The wlan_set_txrx_histogram() API was not returning errors in case of failure.

[Root Cause]
When the HostCmd_CMD_TX_RX_PKT_STATS command fails, wlan_set_txrx_histogram()
was not returning an error because the return value of wifi_set_txrx_histogram()
was not being checked.

[Fix]
Added error checks in the wlan_set_txrx_histogram() API.

[Port Proposal]
Non-general.

[Testing]
Verified that the "wlan-txrx-histogram" CLI now throws an error when the
command fails.

Change-Id: 76fc04c6-b96b-11f0-b249-00155dcb092d

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>