Fix mismatch between GO and STA channels causing GC connection failure.
[Description]
When STA is connected to an AP on channel A and a GO is started on channel B,
driver and wpa_supplicant assume GO is on channel B. However, firmware forces
GO to channel A to sync with STA, causing GC to fail to connect.

[Root Cause]
Channel information between driver/supplicant and firmware is not synchronized
for GO, leading to incorrect assumptions and connection failure.

[Fix]
1. Add a helper function to get frequency from channel.
2. Update p2p_add_group command to replace freq value with the frequency of
   the STA's current channel.

[Port proposal]
non-general

[Testing]
1. Verified GO switches to STA's channel when STA is connected.
2. GC successfully connects to GO.

[Reviewers]
Gannicus Guo, Pankaj Razdan, Sanjay Yadav.

Change-Id: Icdfa9756d9a2bc52752c91f08aef4145dc32689c
Signed-off-by: Qiankun Li <qiankun.li@nxp.com>