Fix can't connect to scanned extAP successfully
[Description]
As sta, could scan the channel 177 extAP, but could not connect to it

[Root Cause]
By default, the region code is 0x00(world wide) for RW610 and IW610, channel 177 should be disabled for this region.
Although region code is valild, the universal_channel info is used for scan but not the region_channel info.
universal_channel includes channel 177, so could scan the channel 177 extAP.
But when connection, would check whether the channel is available, channel info is consistent with region 0x00 and channel 177 is disabled,
so could not connect to this extAP because channel 177 is disabled in such case.

[Fix]
When create channel list, if region code is valid, use region_channel info for scan but not universal_channel info.

[Testing]
Local test ok, the scan channel list consistent with the region code

[Reviewers]
Gannicus Guo, Hui Bai

[Port proposal]
general

Change-Id: 750289b8bb8411f0927b07d5215548c2

Signed-off-by: Qingling Wu qingling.wu@nxp.com