Add enhancement to sequence number for FW commands
[Description]
The sequence number is made from three components: seq_num, bss_num and
bss_type. We were sending seq_num as 0 for all commands. Ideally seq_num should
be incremented for the next command.

[Root Cause]
Currently we were only using the bss_type field and the rest were zero for
commands that are sent to FW.

[Fix]
Added new API wifi_get_cmd_seq_num() which takes the mlan private struct pointer
for the interface and then increments the seq_num, uses bss_num and bss_type for
that interface. In the future if we add a new interface with different bss_num
then we can take advantage of this API.

[Port Proposal]
Non-general.

[Testing]
1. Verified the sequence number is incremented for each subsequent command sent
   to FW.
2. Verified that normal scan, connect and disconnect works.

Change-Id: 9d482708-c060-11f0-aaa0-00155dcb092d

Signed-off-by: Sanjay Yadav <sanjay.yadav@nxp.com>