Fix mismatch between Negotiation GO and STA channels causing GC connection failure.
[Description]
When STA is connected to an AP on channel A and a Negotiation GO is started on channel B,
GC whose channel(frequency) is the same with GO can't connected to GO.

[Root Cause]
Can't start uAP with negotiation channel.

[Fix]
Update p2p_connect command to replace freq value with the frequency of
the STA's current channel.

[Port proposal]
non-general

[Testing]
1. Verified GO switches to STA's channel when STA is connected.
2. GC successfully connects to GO.

[Reviewers]
Gannicus Guo, Pankaj Razdan, Sanjay Yadav.

Change-Id: I472b8c39c499b04e2daefe3542788a307941cdb0
Signed-off-by: Qiankun Li <qiankun.li@nxp.com>