Fix Ex-STA not connecting to uAP post IR
[Description]
Post IR Ex-STA unable to connect to uAP

[Root Cause]
Management packets indication is set and device is fucntioning normally
just before IR, but once IR is called the Management packet indication
is still as was earlier and due to this the same setting is not
happening again, but since in IR f/w is reset so f/w doesnt have the
context of Management packet indication. Hence once IR is done
needed management packets are not sent to driver causing Ex-STA connect
issues.

[Fix]
mgmt_subtype_mask is not cleared during IR, clearing this is ensuring
that managment indication is set again in firmware after every reset

[Testing]
Build pass and tested to behave ok with uAP coming up and Ex-STA
connecting to it

Change-Id: 2839204e-3321-46d6-b5e4-0989be66d040

Signed-off-by: Supriya Inamdar <supriyasumeet.inamdar@nxp.com>